Refactor task and collection packages
All the Task classes are put into "task" while TaskList is put into
"collection".
