+++ src/main/java/tecna/task/Deadline.java
@@ -1,4 +1,4 @@
-package tecna;
+package tecna.task;
 
 import java.time.format.DateTimeFormatter;
 import java.time.LocalDateTime;
