+++ src/main/java/tecna/task/Task.java
@@ -1,4 +1,4 @@
-package tecna;
+package tecna.task;
 
 /**
  * Represents a task entered by the user.
