+++ src/main/java/tecna/ui/Ui.java
@@ -1,7 +1,7 @@
 package tecna.ui;
 
-import tecna.TaskList;
-import tecna.Task;
+import tecna.collection.TaskList;
+import tecna.task.Task;
 
 public class Ui {
     public String printLogo() {
