+++ src/test/java/tecna/StorageTest.java
@@ -17,8 +17,8 @@ public class StorageTest {
         tecna.storage.Storage storage = new tecna.storage.Storage("data/tasklist_not_found.json");
 
         try {
-            ArrayList<Task> tasks = new ArrayList<>();
-            ToDo toDo = new ToDo("go to school", false);
+            ArrayList<tecna.task.Task> tasks = new ArrayList<>();
+            tecna.task.ToDo toDo = new tecna.task.ToDo("go to school", false);
             tasks.add(toDo);
 
             assertEquals(tasks, storage.load());
@@ -33,8 +33,8 @@ public class StorageTest {
     public void load_invalidTaskType_exceptionThrown() {
         tecna.storage.Storage storage = new tecna.storage.Storage("data/error_tecna.json");
         try {
-            ArrayList<Task> tasks = new ArrayList<>();
-            ToDo toDo = new ToDo("borrow a book", true);
+            ArrayList<tecna.task.Task> tasks = new ArrayList<>();
+            tecna.task.ToDo toDo = new tecna.task.ToDo("borrow a book", true);
             tasks.add(toDo);
 
             assertEquals(tasks, storage.load());
@@ -49,9 +49,9 @@ public class StorageTest {
     public void load_deadlineDataMissing_exceptionThrown() {
         tecna.storage.Storage storage = new tecna.storage.Storage("data/deadline_data_missing.json");
         try {
-            ArrayList<Task> tasks = new ArrayList<>();
+            ArrayList<tecna.task.Task> tasks = new ArrayList<>();
             DateTimeFormatter pattern = DateTimeFormatter.ofPattern("yyyy-MM-dd HHmm");
-            Deadline deadline = new Deadline("borrow a book", true, LocalDateTime.parse("2024-02-04 1700", pattern));
+            tecna.task.Deadline deadline = new tecna.task.Deadline("borrow a book", true, LocalDateTime.parse("2024-02-04 1700", pattern));
             tasks.add(deadline);
 
             assertEquals(tasks, storage.load());
